+++ ReleaseNotes/release_notes.md
@@ -9,6 +9,7 @@
   1. add `graph_summary`
   1. update `graph_summary` to add colored bands for "averages" `
   1. replace `ROOT.Double` with `ctypes.c_double`
+  1. `graph_summary` : add labels and type `Graph`
   
   
 ## Backward incompatible changes